only include collection in most recent if it has a minimum of 5% of the votes from the top collection
For example, if the top collection has 100 votes, a collection must have at least 5 and at most 100 votes for it to be considered in the 'most recently curated collections' list.

This way we can scale along as the amount of curators grows and avoid trash collections to be shown in the carousel.